requirements.txt
gradio
tesseract
pillow
